requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gspread==6.1.2
google-auth==2.29.0
pydantic==2.7.1
python-multipart==0.0.9
